# backend/swagger.json
{
	"components": {
...
		"schemas": {
			"TMatchState": {
				"type": "string",
				"enum": [
					"ELECTION",
					"MATCH_MAP",
					"FINISHED"
				],
				"description": "Possible match states."
			},
...
			"IElectionStep": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/IElectionStepAdd"
					},
					{
						"$ref": "#/components/schemas/IElectionStepSkip"
					}
				],
				"description": "Election step to either add map to the list of match maps. Or to remove a map from the map pool."
			},
...
			"TMatchEndAction": {
				"type": "string",
				"enum": [
					"KICK_ALL",
					"QUIT_SERVER",
					"NONE"
				]
			},
...
			"IMatch": {
				"properties": {
					"id": {
						"type": "string",
						"description": "tmt2 identifier for this match"
					},
					"state": {
						"$ref": "#/components/schemas/TMatchState"
					},
					"passthrough": {
						"type": "string",
						"description": "e.g. remote identifier, will be present in every response/webhook"
					},
					"mapPool": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "The maps the players can pick or ban.\nWill also be used if a map is chosen randomly.\nIf the map is fixed it will not be removed from the map pool."
					},
					"teamA": {
						"$ref": "#/components/schemas/ITeam",
						"description": "Team A for this match.\nTeam A will always be Team A in responses and webhooks.\nNo matter on which side (CT/T) it is currently."
					},
					"teamB": {
						"$ref": "#/components/schemas/ITeam",
						"description": "Team B for this match.\nTeam B will always be Team B in responses and webhooks.\nNo matter on which side (CT/T) it is currently."
					},
					"electionSteps": {
						"items": {
							"$ref": "#/components/schemas/IElectionStep"
						},
						"type": "array",
						"description": "List of election steps to determine the played map(s)."
					},
					"election": {
						"$ref": "#/components/schemas/IElection",
						"description": "Data for the election process."
					},
					"gameServer": {
						"$ref": "#/components/schemas/IGameServer"
					},
					"logSecret": {
						"type": "string",
						"description": "Log secret that is given as part of the url to the CS2 server as a log receiver (logaddress_add_http)."
					},
					"parseIncomingLogs": {
						"type": "boolean",
						"description": "Indicates if incoming logs from the CS2 server are parsed (otherwise they will be dropped without any action).\nWill be set to true if match is loaded from storage (after a short delay)."
					},
					"matchMaps": {
						"items": {
							"$ref": "#/components/schemas/IMatchMap"
						},
						"type": "array",
						"description": "The maps which will be played. If match state is still ELECTION than this is not final."
					},
					"currentMap": {
						"type": "number",
						"format": "double",
						"description": "Index of the matchMaps array indicating the current map."
					},
					"webhookUrl": {
						"type": "string",
						"nullable": true,
						"description": "if set various events will be posted to this url"
					},
					"rconCommands": {
						"properties": {
							"end": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed after last match map"
							},
							"match": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every match map start"
							},
							"knife": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every knife round"
							},
							"init": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed exactly once on match init"
							}
						},
						"required": [
							"end",
							"match",
							"knife",
							"init"
						],
						"type": "object"
					},
					"canClinch": {
						"type": "boolean",
						"description": "defaults to true, means that possibly not all maps will be played if the winner is determined before"
					},
					"matchEndAction": {
						"$ref": "#/components/schemas/TMatchEndAction",
						"description": "defaults to NONE"
					},
					"logs": {
						"items": {
							"$ref": "#/components/schemas/TLogUnion"
						},
						"type": "array"
					},
					"players": {
						"items": {
							"$ref": "#/components/schemas/IPlayer"
						},
						"type": "array"
					},
					"tmtSecret": {
						"type": "string",
						"description": "Access token to be used in the API."
					},
					"isStopped": {
						"type": "boolean",
						"description": "If match is finished or if the match was stopped/deleted this is true."
					},
					"serverPassword": {
						"type": "string",
						"description": "Server password, periodically fetched from game server"
					},
					"tmtLogAddress": {
						"type": "string",
						"description": "if set will be used to register the target logaddress for the game server"
					},
					"createdAt": {
						"type": "number",
						"format": "double",
						"description": "Creation date (unix time in milliseconds since midnight, January 1, 1970 UTC)"
					},
					"mode": {
						"$ref": "#/components/schemas/TMatchMode",
						"description": "Match mode (single: stops when match is finished, loop: starts again after match is finished)"
					}
				},
				"required": [
					"id",
					"state",
					"mapPool",
					"teamA",
					"teamB",
					"electionSteps",
					"election",
					"gameServer",
					"logSecret",
					"parseIncomingLogs",
					"matchMaps",
					"currentMap",
					"webhookUrl",
					"rconCommands",
					"canClinch",
					"matchEndAction",
					"logs",
					"players",
					"tmtSecret",
					"isStopped",
					"serverPassword",
					"createdAt",
					"mode"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IMatchCreateDto": {
				"properties": {
					"passthrough": {
						"type": "string",
						"description": "e.g. remote identifier, will be present in every response/webhook"
					},
					"mapPool": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "The maps the players can pick or ban.\nWill also be used if a map is chosen randomly.\nIf the map is fixed it will not be removed from the map pool."
					},
					"teamA": {
						"$ref": "#/components/schemas/ITeamCreateDto"
					},
					"teamB": {
						"$ref": "#/components/schemas/ITeamCreateDto"
					},
					"electionSteps": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/IElectionStepAdd"
								},
								{
									"$ref": "#/components/schemas/IElectionStepSkip"
								}
							]
						},
						"type": "array"
					},
					"gameServer": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IGameServer"
							}
						],
						"nullable": true
					},
					"webhookUrl": {
						"type": "string",
						"nullable": true,
						"description": "if set various events will be posted to this url"
					},
					"rconCommands": {
						"properties": {
							"end": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed after last match map"
							},
							"match": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every match map start"
							},
							"knife": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every knife round"
							},
							"init": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed exactly once on match init"
							}
						},
						"type": "object"
					},
					"canClinch": {
						"type": "boolean",
						"description": "defaults to true, means that possibly not all maps will be played if the winner is determined before"
					},
					"matchEndAction": {
						"$ref": "#/components/schemas/TMatchEndAction",
						"description": "defaults to NONE"
					},
					"tmtLogAddress": {
						"type": "string",
						"description": "if set will be used to register the target logaddress for the game server"
					},
					"mode": {
						"$ref": "#/components/schemas/TMatchMode",
						"description": "Match mode (single: stops when match is finished, loop: starts again after match is finished)"
					}
				},
				"required": [
					"mapPool",
					"teamA",
					"teamB",
					"electionSteps",
					"gameServer"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IMatchResponse": {
				"properties": {
					"id": {
						"type": "string",
						"description": "tmt2 identifier for this match"
					},
					"state": {
						"$ref": "#/components/schemas/TMatchState"
					},
					"passthrough": {
						"type": "string",
						"description": "e.g. remote identifier, will be present in every response/webhook"
					},
					"mapPool": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "The maps the players can pick or ban.\nWill also be used if a map is chosen randomly.\nIf the map is fixed it will not be removed from the map pool."
					},
					"teamA": {
						"$ref": "#/components/schemas/ITeam",
						"description": "Team A for this match.\nTeam A will always be Team A in responses and webhooks.\nNo matter on which side (CT/T) it is currently."
					},
					"teamB": {
						"$ref": "#/components/schemas/ITeam",
						"description": "Team B for this match.\nTeam B will always be Team B in responses and webhooks.\nNo matter on which side (CT/T) it is currently."
					},
					"electionSteps": {
						"items": {
							"$ref": "#/components/schemas/IElectionStep"
						},
						"type": "array",
						"description": "List of election steps to determine the played map(s)."
					},
					"election": {
						"$ref": "#/components/schemas/IElection",
						"description": "Data for the election process."
					},
					"gameServer": {
						"$ref": "#/components/schemas/IGameServer"
					},
					"logSecret": {
						"type": "string",
						"description": "Log secret that is given as part of the url to the CS2 server as a log receiver (logaddress_add_http)."
					},
					"parseIncomingLogs": {
						"type": "boolean",
						"description": "Indicates if incoming logs from the CS2 server are parsed (otherwise they will be dropped without any action).\nWill be set to true if match is loaded from storage (after a short delay)."
					},
					"matchMaps": {
						"items": {
							"$ref": "#/components/schemas/IMatchMap"
						},
						"type": "array",
						"description": "The maps which will be played. If match state is still ELECTION than this is not final."
					},
					"currentMap": {
						"type": "number",
						"format": "double",
						"description": "Index of the matchMaps array indicating the current map."
					},
					"webhookUrl": {
						"type": "string",
						"nullable": true,
						"description": "if set various events will be posted to this url"
					},
					"rconCommands": {
						"properties": {
							"end": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed after last match map"
							},
							"match": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every match map start"
							},
							"knife": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every knife round"
							},
							"init": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed exactly once on match init"
							}
						},
						"required": [
							"end",
							"match",
							"knife",
							"init"
						],
						"type": "object"
					},
					"canClinch": {
						"type": "boolean",
						"description": "defaults to true, means that possibly not all maps will be played if the winner is determined before"
					},
					"matchEndAction": {
						"$ref": "#/components/schemas/TMatchEndAction",
						"description": "defaults to NONE"
					},
					"logs": {
						"items": {
							"$ref": "#/components/schemas/TLogUnion"
						},
						"type": "array"
					},
					"players": {
						"items": {
							"$ref": "#/components/schemas/IPlayer"
						},
						"type": "array"
					},
					"tmtSecret": {
						"type": "string",
						"description": "Access token to be used in the API."
					},
					"isStopped": {
						"type": "boolean",
						"description": "If match is finished or if the match was stopped/deleted this is true."
					},
					"serverPassword": {
						"type": "string",
						"description": "Server password, periodically fetched from game server"
					},
					"tmtLogAddress": {
						"type": "string",
						"description": "if set will be used to register the target logaddress for the game server"
					},
					"createdAt": {
						"type": "number",
						"format": "double",
						"description": "Creation date (unix time in milliseconds since midnight, January 1, 1970 UTC)"
					},
					"mode": {
						"$ref": "#/components/schemas/TMatchMode",
						"description": "Match mode (single: stops when match is finished, loop: starts again after match is finished)"
					},
					"isLive": {
						"type": "boolean"
					}
				},
				"required": [
					"id",
					"state",
					"mapPool",
					"teamA",
					"teamB",
					"electionSteps",
					"election",
					"gameServer",
					"logSecret",
					"parseIncomingLogs",
					"matchMaps",
					"currentMap",
					"webhookUrl",
					"rconCommands",
					"canClinch",
					"matchEndAction",
					"logs",
					"players",
					"tmtSecret",
					"isStopped",
					"serverPassword",
					"createdAt",
					"mode",
					"isLive"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Event": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/ChatEvent"
					},
					{
						"$ref": "#/components/schemas/ElectionEndEvent"
					},
					{
						"$ref": "#/components/schemas/RoundEndEvent"
					},
					{
						"$ref": "#/components/schemas/MapEndEvent"
					},
					{
						"$ref": "#/components/schemas/MatchEndEvent"
					},
					{
						"$ref": "#/components/schemas/KnifeRoundEndEvent"
					},
					{
						"$ref": "#/components/schemas/MapStartEvent"
					},
					{
						"$ref": "#/components/schemas/LogEvent"
					},
					{
						"$ref": "#/components/schemas/ElectionMapStep"
					},
					{
						"$ref": "#/components/schemas/ElectionSideStep"
					},
					{
						"$ref": "#/components/schemas/MatchCreateEvent"
					},
					{
						"$ref": "#/components/schemas/MatchUpdateEvent"
					}
				]
			},
			"IMatchUpdateDto": {
				"properties": {
					"passthrough": {
						"type": "string",
						"description": "e.g. remote identifier, will be present in every response/webhook"
					},
					"mapPool": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "The maps the players can pick or ban.\nWill also be used if a map is chosen randomly.\nIf the map is fixed it will not be removed from the map pool."
					},
					"teamA": {
						"$ref": "#/components/schemas/ITeamCreateDto"
					},
					"teamB": {
						"$ref": "#/components/schemas/ITeamCreateDto"
					},
					"electionSteps": {
						"items": {
							"anyOf": [
								{
									"$ref": "#/components/schemas/IElectionStepAdd"
								},
								{
									"$ref": "#/components/schemas/IElectionStepSkip"
								}
							]
						},
						"type": "array"
					},
					"gameServer": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IGameServer"
							}
						],
						"nullable": true
					},
					"webhookUrl": {
						"type": "string",
						"nullable": true,
						"description": "if set various events will be posted to this url"
					},
					"rconCommands": {
						"properties": {
							"end": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed after last match map"
							},
							"match": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every match map start"
							},
							"knife": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed before every knife round"
							},
							"init": {
								"items": {
									"type": "string"
								},
								"type": "array",
								"description": "executed exactly once on match init"
							}
						},
						"type": "object"
					},
					"canClinch": {
						"type": "boolean",
						"description": "defaults to true, means that possibly not all maps will be played if the winner is determined before"
					},
					"matchEndAction": {
						"$ref": "#/components/schemas/TMatchEndAction",
						"description": "defaults to NONE"
					},
					"tmtLogAddress": {
						"type": "string",
						"description": "if set will be used to register the target logaddress for the game server"
					},
					"mode": {
						"$ref": "#/components/schemas/TMatchMode",
						"description": "Match mode (single: stops when match is finished, loop: starts again after match is finished)"
					},
					"state": {
						"$ref": "#/components/schemas/TMatchState"
					},
					"logSecret": {
						"type": "string",
						"description": "updates the server's log address automatically"
					},
					"currentMap": {
						"type": "number",
						"format": "double"
					},
					"_restartElection": {
						"type": "boolean"
					},
					"_init": {
						"type": "boolean"
					},
					"_setup": {
						"type": "boolean"
					},
					"_execRconCommandsInit": {
						"type": "boolean"
					},
					"_execRconCommandsKnife": {
						"type": "boolean"
					},
					"_execRconCommandsMatch": {
						"type": "boolean"
					},
					"_execRconCommandsEnd": {
						"type": "boolean"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearer_token": {
				"type": "http",
				"scheme": "bearer"
			},
			"bearer_token_optional": {
				"type": "http",
				"scheme": "bearer"
			}
		}
	},
	"info": {
		"title": "tmt2-backend",
		"version": "2.0.0",
		"license": {
			"name": "MIT"
		},
		"contact": {
			"name": "Jens Forstmann"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/api/matches": {
			"post": {
				"operationId": "CreateMatch",
				"responses": {
					"201": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IMatch"
								}
							}
						}
					}
				},
				"description": "Create and supervise a new match.",
				"security": [
					{
						"bearer_token_optional": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IMatchCreateDto"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetAllMatches",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/IMatchResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get all matches.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"description": "State filter",
						"in": "query",
						"name": "state",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"description": "Passthrough filter",
						"in": "query",
						"name": "passthrough",
						"required": false,
						"schema": {
							"type": "array",
							"items": {
								"type": "string"
							}
						}
					},
					{
						"description": "Get only stopped or not stopped matches.",
						"in": "query",
						"name": "isStopped",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					},
					{
						"description": "Filter for only live (currently active) matches, or the opposite.",
						"in": "query",
						"name": "isLive",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			}
		},
		"/api/matches/{id}": {
			"get": {
				"operationId": "GetMatch",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/IMatchResponse"
										},
										{}
									]
								}
							}
						}
					}
				},
				"description": "Get a specific match by id.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateMatch",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Update a specific match.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IMatchUpdateDto"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteMatch",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Stop supervising a specific match. TMT will no longer listen to the game server and will not execute any rcon commands.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/matches/{id}/logs": {
			"get": {
				"operationId": "GetLogs",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"type": "string"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get the last 1000 log lines from a specific match.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/matches/{id}/events": {
			"get": {
				"operationId": "GetEvents",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/Event"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"description": "Get the last 1000 events from a specific match.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/matches/{id}/server/round_backups": {
			"get": {
				"operationId": "GetRoundBackups",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"properties": {
												"total": {
													"type": "number",
													"format": "double"
												},
												"latestFiles": {
													"items": {
														"type": "string"
													},
													"type": "array"
												}
											},
											"required": [
												"total",
												"latestFiles"
											],
											"type": "object"
										},
										{}
									]
								}
							}
						}
					}
				},
				"description": "Get the last known round backups for a specific match.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "The max. number of round backups to be returned.",
						"in": "query",
						"name": "count",
						"required": false,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/api/matches/{id}/server/round_backups/{file}": {
			"post": {
				"operationId": "LoadRoundBackup",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"type": "boolean"
										},
										{}
									]
								}
							}
						}
					}
				},
				"description": "Load a round backup file for a specific match.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Name of the round backup file.",
						"in": "path",
						"name": "file",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/matches/{id}/matchMap/{mapNumber}": {
			"patch": {
				"operationId": "UpdateMatchMap",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Update a specific match map. First map has the map number 0.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "mapNumber",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/IMatchMapUpdateDto"
							}
						}
					}
				}
			}
		},
		"/api/matches/{id}/revive": {
			"patch": {
				"operationId": "ReviveMatch",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Revive a specific match. TMT will start supervising a (stopped) match again (listen to the game sever and execute rcon commands).",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/matches/{id}/server/rcon": {
			"post": {
				"operationId": "Rcon",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"items": {
												"type": "string"
											},
											"type": "array"
										},
										{}
									]
								}
							}
						}
					}
				},
				"description": "Execute a rcon command on the game server.",
				"security": [
					{
						"bearer_token": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"type": "string"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
		"/api/matches/{id}/server/log/{secret}": {
			"post": {
				"operationId": "ReceiveLog",
				"responses": {
					"204": {
						"description": "No content"
					}
				},
				"description": "Endpoint the game server sends its log file to. Not meant for direct use!",
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "path",
						"name": "secret",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}